requirements.txt
flask
python-dotenv
pyTelegramBotAPI